refactor: align project structure with best practices
- Fix deploy-docs.yml trigger path (scripts/convert_xml_to_markdown.js
  → scripts/docs/convert_xml_to_markdown.js)
- Fix deploy-docs.yml Doxyfile reference (Doxyfile → docs/Doxyfile)
- Remove malformed .gitmodules entry [submodule "deps.llhttp"]
- Update Doxyfile PROJECT_NUMBER 2.2.0 → 2.5.0
- Move Doxyfile to docs/Doxyfile (belongs with docs tooling)
- Update .gitignore: add build_asan/ubsan/bench/coverage, remove stale
- Update GNUmakefile docs-clean: stale paths, missing .doxygen/
- Update scripts/README.md: remove references to deleted scripts
- Remove duplicate npm scripts from docs/package.json (dev/docs:dev, etc.)
- Fix fast-xml-parser version to actually resolved version
- Remove pre-generated test files from public/ (benchmark generates them)
- Move scripts/ci/notify_pr.js → .github/scripts/notify_pr.js
- Update scripts/build_docs.sh: Doxyfile → docs/Doxyfile
