+++ .github/workflows/deploy-docs.yml
@@ -6,8 +6,8 @@ on:
     branches: [ main, 'release/**' ]
     paths:
       - 'docs/**'
-      - 'Doxyfile'
-      - 'scripts/convert_xml_to_markdown.js'
+      - 'docs/Doxyfile'
+      - 'scripts/docs/convert_xml_to_markdown.js'
       - '.github/workflows/deploy-docs.yml'
   workflow_dispatch:
 
@@ -47,7 +47,7 @@ jobs:
           mkdir -p docs/api/html docs/api/latex docs/api/xml docs/api/.doxygen
           
           # Generate Doxygen API documentation
-          doxygen Doxyfile
+          doxygen docs/Doxyfile
           
           # Generate API documentation from XML
           cd docs
