+++ .gitmodules
@@ -27,5 +27,3 @@
 [submodule "deps/zlib"]
 	path = deps/zlib
 	url = https://github.com/madler/zlib.git
-[submodule "deps.llhttp"]
-	branch = release
