+++ docs/package.json
@@ -1,1 +1,20 @@
-{"name": "uvhttp-docs", "version": "2.5.0", "type": "module", "scripts": {"dev": "vitepress dev", "build": "vitepress build", "preview": "vitepress preview", "docs:dev": "vitepress dev", "docs:build": "vitepress build", "docs:preview": "vitepress preview", "api:generate": "cd .. && NODE_PATH=docs/node_modules node scripts/docs/convert_xml_to_markdown.js docs/api/.doxygen/xml docs/api/generated", "api:dev": "npm run api:generate && vitepress dev"}, "dependencies": {"fast-xml-parser": "^5.9.0", "nunjucks": "^3.2.4", "xml2js": "^0.6.2"}, "devDependencies": {"vitepress": "^1.6.4"}}
+{
+  "name": "uvhttp-docs",
+  "version": "2.5.0",
+  "type": "module",
+  "scripts": {
+    "docs:dev": "vitepress dev",
+    "docs:build": "vitepress build",
+    "docs:preview": "vitepress preview",
+    "api:generate": "cd .. && NODE_PATH=docs/node_modules node scripts/docs/convert_xml_to_markdown.js docs/api/.doxygen/xml docs/api/generated",
+    "api:dev": "npm run api:generate && vitepress dev"
+  },
+  "dependencies": {
+    "fast-xml-parser": "^5.10.1",
+    "nunjucks": "^3.2.4",
+    "xml2js": "^0.6.2"
+  },
+  "devDependencies": {
+    "vitepress": "^1.6.4"
+  }
+}
